Custom input url modified
