+++ inginious/frontend/plugins/multilang/__init__.py
@@ -1,84 +1,14 @@
 import os
-import web
-import json
 
-from inginious.frontend.plugins.utils import create_static_resource_page, get_mandatory_parameter
-from inginious.client.client_sync import ClientSync
-from inginious.frontend.pages.api._api_page import APIAuthenticatedPage
-from inginious.frontend.parsable_text import ParsableText
+from inginious.frontend.plugins.utils import create_static_resource_page
 from .problems.code_multiple_languages_problem import DisplayableCodeMultipleLanguagesProblem
 from .problems.code_multiple_file_languages_problem import DisplayableCodeFileMultipleLanguagesProblem
 from .problems.notebook_file_problem import DisplayableNotebookFileProblem
 from .problems.constants import set_linter_url, set_python_tutor_url, set_show_tools, get_show_tools
-from bson.objectid import ObjectId
+from .problems.custom_input import custom_input_manager_with_curried_client
 
 _static_folder_path = os.path.join(os.path.dirname(__file__), "static")
 
-def customInputManagerWithCurriedClient(client):
-    class CustomInputManager(APIAuthenticatedPage):
-        def __init__(self):
-            self._client = client
-
-        def add_unsaved_job(self, task, inputdata):
-            temp_client = ClientSync(self._client)
-            return temp_client.new_job(task, inputdata)
-
-        def API_GET(self):
-            request_params = web.input()
-            submission_id = get_mandatory_parameter(request_params, "submission")
-
-            # A mongo request for the submission and his input
-            # How do i do this?
-            things = self.database.submissions.aggregate([
-                {
-                    '$match': {
-                        "_id": ObjectId(submission_id)
-                    }
-                },
-                {
-                    "$project": {
-                        "_id": 0,
-                        "taskid": 1,
-                        "courseid": 1
-                    }
-                }
-            ])
-            return 200, list(things)[0]
-
-        def API_POST(self):
-            request_params = web.input()
-
-            courseid = get_mandatory_parameter(request_params, "courseid")
-            course = self.course_factory.get_course(courseid)
-            taskid = get_mandatory_parameter(request_params, "taskid")
-            task = self.task_factory.get_task(course, taskid)
-
-            try:
-                init_var = {
-                    problem.get_id(): problem.input_type()()
-                    for problem in task.get_problems() if problem.input_type() in [dict, list]
-                }
-                user_input = task.adapt_input_for_backend(web.input(**init_var))
-                for key, value in user_input.items():
-                    if type(value) is str:
-                        user_input[key] = user_input[key].replace("\r", "")
-                result, grade, problems, tests, custom, archive, stdout, stderr = self.add_unsaved_job(task, user_input)
-
-                data = {
-                    "status": ("done" if result[0] == "success" or result[0] == "failed" else "error"),
-                    "result": result[0],
-                    "text": ParsableText(result[1]).parse(),
-                    "stdout": custom.get("custom_stdout", ""),
-                    "stderr": custom.get("custom_stderr", "")
-                }
-                web.header('Content-Type', 'application/json')
-                return 200, json.dumps(data)
-
-            except Exception as ex:
-                web.header('Content-Type', 'application/json')
-                return 200, json.dumps({"status": "error", "text": str(ex)})
-
-    return CustomInputManager
 
 def init(plugin_manager, course_factory, client, plugin_config):
     # This option is to hide/show tools like Python tutor.
@@ -88,7 +18,7 @@ def init(plugin_manager, course_factory, client, plugin_config):
 
     plugin_manager.add_page(r'/multilang/static/(.*)', create_static_resource_page(_static_folder_path))
 
-    plugin_manager.add_page("/api/multilang/", customInputManagerWithCurriedClient(client))
+    plugin_manager.add_page("/api/custom_input/", custom_input_manager_with_curried_client(client))
 
     use_minified = plugin_config.get("use_minified", True)
